refactor(database): reconcile duplicate TaskCompletion entity definitions
