+++ src/users/users.module.ts
@@ -3,7 +3,7 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { UsersController } from './users.controller';
 import { UsersService } from './users.service';
 import { User } from '../entities/user.entity';
-import { TaskCompletion } from './entities/task-completion.entity';
+import { TaskCompletion } from '../tasks/entities/task-completion.entity';
 import { Coupon } from './entities/coupon.entity';
 import { WalletModule } from './wallet/wallet.module';
